+++ packages/core/assets/onload.ts
@@ -26,7 +26,7 @@ if ( __PLUGINS ) {
         return { id, load }
     })
 
-    sanitized.forEach(({ id, load }) => {
+    sanitized.forEach(async ({ id, load }) => {
         
         loaded[id] = undefined // Register that all supported plugins are technically loaded
 
@@ -43,17 +43,19 @@ if ( __PLUGINS ) {
                 } : {}
 
                 loaded[id] = load.call(ctx, ENV)
+                await loaded[id]
+                if (DESKTOP) TEMP_COMMONERS.send(["commoners:loaded", DESKTOP.__id, id].join(":")) // Notify the main process that the plugin is loaded
             }
 
         } catch (e) {
             pluginErrorMessage(id, "load", e)
         }
-
     })
 
     ENV.PLUGINS = loaded
-
     __READY(loaded)
 })
 
-} else __READY({})
+} else {
+    __READY({})
+}